Create build_and_test_app.yml (#10)
From now on, the app is part of the checks in github actions. 

Note:
There is a dirty fix to pass the tests. The issue is due to usage of older version of jest. Jest can't be updated because the older version is dependent by react-create-script.
The proper solution is to create the app from scratch without using react-create-script